Gson, OkHttpClient 분리 가독성 개선
