+++ app/src/main/java/com/example/sunflower/network/GsonConverter.kt
@@ -1,0 +1,7 @@
+package com.example.sunflower.network
+
+import retrofit2.converter.gson.GsonConverterFactory
+
+object GsonConverter {
+    val gSonConverter: GsonConverterFactory = GsonConverterFactory.create()
+}
