+++ app/src/main/java/com/example/sunflower/network/OkHttpClient.kt
@@ -1,0 +1,14 @@
+package com.example.sunflower.network
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+
+object OkHttpClient {
+    private val logger = HttpLoggingInterceptor().apply {
+        level = HttpLoggingInterceptor.Level.BASIC
+    }
+
+    val client = OkHttpClient.Builder()
+        .addInterceptor(logger)
+        .build()
+}
